+++ src/components/Profile/Profile.css
@@ -50,7 +50,7 @@
   gap: 16px;
 }
 
-.profile__bottom-button {
+.profile__edit-button {
   border: none;
   background-color: transparent;
   font-size: 13px;
@@ -60,26 +60,33 @@
   margin: 0;
 }
 
-.profile__bottom-button:disabled {
-  color: gray;
-  opacity: 0.2;
-}
-
-.profile__bottom-button:disabled:hover {
-  color: gray;
-  opacity: 0.2;
-  cursor:default;
-}
-
-
-.profile__bottom-button_color_red {
+.profile__exit-button {
+  border: none;
+  background-color: transparent;
+  font-size: 13px;
+  line-height: 16px;
+  text-align: center;
+  padding: 0;
+  margin: 0;
   color: #ee3465;
 }
 
-.profile__bottom-button:hover {
+.profile__exit-button:hover {
   opacity: .7;
   cursor: pointer;
 }
+
+.profile__edit-button:hover {
+  opacity: .7;
+  cursor: pointer;
+}
+
+.profile__edit-button:disabled {
+  color: red;
+  opacity: 1;
+  cursor: default;
+}
+
 
 @media screen and (max-width: 769px) {
   .profile {
@@ -111,7 +118,12 @@
     margin-top: 379px;
   }
 
-  .profile__bottom-button {
+  .profile__exit-button {
+    font-size: 12px;
+    line-height: 15px;
+  }
+
+  .profile__edit-button {
     font-size: 12px;
     line-height: 15px;
   }
