+++ src/components/SavedMovies/SavedMovies.jsx
@@ -4,22 +4,20 @@ import { useState } from "react";
 import { useEffect } from "react";
 function SavedMovies({ cards, handleDeleteMovie, isLoading, savedMovies }) {
   const [filteredCards, setFilteredCards] = useState(cards);
-  const [notShortFilms, setNotShortFilms] = useState(cards)
+  const [notShortFilms, setNotShortFilms] = useState(cards);
   const [isShorts, setShorts] = useState(false);
-  
-  useEffect(()=>{
+
+  useEffect(() => {
     const likedCards = cards.filter((card) => card.isLiked);
-    setFilteredCards(likedCards)    
-  },[savedMovies, cards])
+    setFilteredCards(likedCards);
+  }, [savedMovies, cards]);
 
-
-  useEffect(()=>{
-    const shorts = filteredCards.filter((elem)=> {
+  useEffect(() => {
+    const shorts = filteredCards.filter((elem) => {
       return elem.duration > 40;
-    })
+    });
     setNotShortFilms(shorts);
-  },[filteredCards, savedMovies])
-
+  }, [filteredCards, savedMovies]);
 
   function search(e) {
     e.preventDefault();
@@ -32,24 +30,29 @@ function SavedMovies({ cards, handleDeleteMovie, isLoading, savedMovies }) {
 
   function filterCards(keywords) {
     setShorts(JSON.parse(localStorage.getItem("shorts")));
-    
 
     const filteredCards = cards.filter((card) => {
-      if(!card.isLiked){
+      if (!card.isLiked) {
         return false;
       }
       return card.nameRU.toLowerCase().includes(keywords.toLowerCase());
     });
     setFilteredCards(filteredCards);
-    setNotShortFilms(filteredCards.filter((elem)=> {
-      return elem.duration > 40;
-    }));
+    setNotShortFilms(
+      filteredCards.filter((elem) => {
+        return elem.duration > 40;
+      })
+    );
   }
 
   return (
     <section className="saved-movies">
-      <SearchForm handleSearch={search}/>
-      <MoviesCardList cards={!isShorts ? notShortFilms : filteredCards} handleDeleteMovie={handleDeleteMovie} isLoading={isLoading}/>
+      <SearchForm handleSearch={search} />
+      <MoviesCardList
+        cards={!isShorts ? notShortFilms : filteredCards}
+        handleDeleteMovie={handleDeleteMovie}
+        isLoading={isLoading}
+      />
     </section>
   );
 }
