+++ src/components/SearchForm/SearchForm.jsx
@@ -1,34 +1,33 @@
 import { useState } from "react";
 import "./SearchForm.css";
 
-function SearchForm({handleSearch}) {
-  const [keyword, setKeyword] = useState('');
+function SearchForm({ handleSearch }) {
+  const [keyword, setKeyword] = useState("");
   const [showShort, setShowShort] = useState(false);
 
-  useState(()=>{
-    if(localStorage.getItem("keywords")){
+  useState(() => {
+    if (localStorage.getItem("keywords")) {
       const keyword = localStorage.getItem("keywords");
       setKeyword(keyword);
     }
-    if(localStorage.getItem("shorts")){
+    if (localStorage.getItem("shorts")) {
       const shorts = JSON.parse(localStorage.getItem("shorts"));
       setShowShort(shorts);
     } else {
-      localStorage.setItem("shorts", JSON.stringify(showShort))
+      localStorage.setItem("shorts", JSON.stringify(showShort));
     }
-  },[])
+  }, []);
 
-  function handleChange(e){
+  function handleChange(e) {
     setKeyword(e.target.value);
-    localStorage.setItem("keywords", e.target.value)
+    localStorage.setItem("keywords", e.target.value);
   }
 
-  function toggleShort(e){
-    setShowShort(e.target.checked)
-    localStorage.setItem("shorts", e.target.checked)
+  function toggleShort(e) {
+    setShowShort(e.target.checked);
+    localStorage.setItem("shorts", e.target.checked);
   }
 
-  
   return (
     <form className="search-form">
       <div className="search-form__input-container">
@@ -41,7 +40,11 @@ function SearchForm({handleSearch}) {
           value={keyword}
           onChange={handleChange}
         />
-        <button type="submit" className="search-form__button" onClick={handleSearch}>
+        <button
+          type="submit"
+          className="search-form__button"
+          onClick={handleSearch}
+        >
           Найти
         </button>
       </div>
@@ -52,7 +55,6 @@ function SearchForm({handleSearch}) {
           className="search-form__invisible-checkbox"
           checked={showShort}
           onChange={toggleShort}
-          
         />
         <div className="search-form__visible-checkbox-frame">
           <div className="search-form__visible-checkbox-inner-circle"></div>
