+++ src/components/Techs/Techs.jsx
