--- src/utils/MoviesApiConfig.js
@@ -1,9 +1,0 @@
-
-const MainApiConfig = {
-    commonUrlPart: 'https://api.nomoreparties.co/beatfilm-movies',
-    headers: {
-        'Content-Type': 'application/json',
-        Authorization: 'Bearer ' + localStorage.getItem("jwt"),
-      }
-}
-export { MainApiConfig }
